+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/measurements/tests/Test_MPSTNCuda_Expval.cpp
@@ -35,23 +35,23 @@ using namespace Pennylane::LightningTensor::TNCuda::Observables;
 /// @endcond
 
 TEMPLATE_TEST_CASE("[Identity]", "[MPSTNCuda_Expval]", float, double) {
-    using StateTensorT = MPSTNCuda<TestType>;
-    using NamedObsT = NamedObsTNCuda<StateTensorT>;
+    using TensorNetT = MPSTNCuda<TestType>;
+    using NamedObsT = NamedObsTNCuda<TensorNetT>;
     auto ONE = TestType(1);
 
     std::size_t bondDim = GENERATE(2, 3, 4, 5);
     std::size_t num_qubits = 3;
     std::size_t maxBondDim = bondDim;
 
-    StateTensorT mps_state{num_qubits, maxBondDim};
+    TensorNetT mps_state{num_qubits, maxBondDim};
 
-    auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+    auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
 
     SECTION("Using expval") {
         mps_state.applyOperations({{"Hadamard"}, {"CNOT"}, {"CNOT"}},
                                   {{0}, {0, 1}, {1, 2}},
                                   {{false}, {false}, {false}});
-        mps_state.get_final_state();
+        mps_state.append_mps_final_state();
         auto ob = NamedObsT("Identity", {0});
         auto res = measure.expval(ob);
         CHECK(res == Approx(ONE));
@@ -60,16 +60,16 @@ TEMPLATE_TEST_CASE("[Identity]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("[PauliX]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using NamedObsT = NamedObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using NamedObsT = NamedObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
-        auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto ZERO = TestType(0);
         auto ONE = TestType(1);
@@ -78,7 +78,7 @@ TEMPLATE_TEST_CASE("[PauliX]", "[MPSTNCuda_Expval]", float, double) {
             mps_state.applyOperations({{"Hadamard"}, {"CNOT"}, {"CNOT"}},
                                       {{0}, {0, 1}, {1, 2}},
                                       {{false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = NamedObsT("PauliX", {0});
             auto res = measure.expval(ob);
             CHECK(res == ZERO);
@@ -88,7 +88,7 @@ TEMPLATE_TEST_CASE("[PauliX]", "[MPSTNCuda_Expval]", float, double) {
             mps_state.applyOperations(
                 {{"Hadamard"}, {"Hadamard"}, {"Hadamard"}}, {{0}, {1}, {2}},
                 {{false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = NamedObsT("PauliX", {0});
             auto res = measure.expval(ob);
             CHECK(res == Approx(ONE));
@@ -104,7 +104,7 @@ TEMPLATE_TEST_CASE("[PauliX]", "[MPSTNCuda_Expval]", float, double) {
                  {"Hadamard"}},
                 {{0}, {0}, {1}, {1}, {2}, {2}},
                 {{false}, {false}, {false}, {false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = NamedObsT("PauliX", {0});
             auto res = measure.expval(ob);
             CHECK(res == -Approx(ONE));
@@ -114,16 +114,16 @@ TEMPLATE_TEST_CASE("[PauliX]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("[PauliY]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using NamedObsT = NamedObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using NamedObsT = NamedObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
-        auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto ZERO = TestType(0);
         auto ONE = TestType(1);
@@ -160,42 +160,66 @@ TEMPLATE_TEST_CASE("[PauliY]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("[PauliZ]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using PrecisionT = StateTensorT::PrecisionT;
-        using StateTensorT = MPSTNCuda<TestType>;
-        using NamedObsT = NamedObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using PrecisionT = TensorNetT::PrecisionT;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using NamedObsT = NamedObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
         SECTION("Using expval") {
             mps_state.applyOperations(
                 {{"RX"}, {"Hadamard"}, {"Hadamard"}}, {{0}, {1}, {2}},
                 {{false}, {false}, {false}}, {{0.5}, {}, {}});
-            auto m = MeasurementsTNCuda<StateTensorT>(mps_state);
+            auto m = MeasurementsTNCuda<TensorNetT>(mps_state);
             auto ob = NamedObsT("PauliZ", {0});
             auto res = m.expval(ob);
             PrecisionT ref = 0.8775825618903724;
             REQUIRE(res == Approx(ref).margin(1e-6));
+        }
+
+        SECTION("Using expval mps with cutoff") {
+            double cutoff = GENERATE(1e-1, 1e-2);
+            std::string cutoff_mode = GENERATE("rel", "abs");
+            mps_state.applyOperations(
+                {{"Hadamard"},
+                 {"Hadamard"},
+                 {"Hadamard"},
+                 {"SingleExcitation"},
+                 {"IsingXX"},
+                 {"IsingXY"}},
+                {{0}, {1}, {2}, {0, 1}, {1, 2}, {0, 2}},
+                {{false}, {false}, {false}, {false}, {false}, {false}},
+                {{}, {}, {}, {0.5}, {0.6}, {0.7}});
+            mps_state.append_mps_final_state(cutoff, cutoff_mode);
+            auto m = MeasurementsTNCuda<TensorNetT>(mps_state);
+            auto ob = NamedObsT("PauliZ", {0});
+            auto res = m.expval(ob);
+            // ref is from default.qubit
+            PrecisionT ref = -0.2115276040475712;
+
+            REQUIRE_THAT(res, Catch::Matchers::WithinRel(
+                                  ref, static_cast<PrecisionT>(cutoff)));
         }
     }
 }
 
 TEMPLATE_TEST_CASE("[Hadamard]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using NamedObsT = NamedObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using NamedObsT = NamedObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
-        auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto INVSQRT2 = TestType(0.707106781186547524401);
 
@@ -205,7 +229,7 @@ TEMPLATE_TEST_CASE("[Hadamard]", "[MPSTNCuda_Expval]", float, double) {
         // multiple times with different observables
         SECTION("Using expval") {
             mps_state.applyOperation("PauliX", {0});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
 
             auto ob = NamedObsT("Hadamard", {0});
             auto res = measure.expval(ob);
@@ -220,21 +244,21 @@ TEMPLATE_TEST_CASE("[Hadamard]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("[Parametric_obs]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using NamedObsT = NamedObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using NamedObsT = NamedObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
-        auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
         auto ONE = TestType(1);
 
         SECTION("Using expval") {
             mps_state.applyOperation("PauliX", {0});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
 
             auto ob = NamedObsT("RX", {0}, {0});
             auto res = measure.expval(ob);
@@ -245,17 +269,17 @@ TEMPLATE_TEST_CASE("[Parametric_obs]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("[Hermitian]", "[MPSTNCuda_Expval]", float, double) {
     {
-        using StateTensorT = MPSTNCuda<TestType>;
-        using ComplexT = typename StateTensorT::ComplexT;
-        using HermitianObsT = HermitianObsTNCuda<StateTensorT>;
+        using TensorNetT = MPSTNCuda<TestType>;
+        using ComplexT = typename TensorNetT::ComplexT;
+        using HermitianObsT = HermitianObsTNCuda<TensorNetT>;
 
         std::size_t bondDim = GENERATE(2, 3, 4, 5);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
-        auto measure = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto measure = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto ZERO = TestType(0);
         auto ONE = TestType(1);
@@ -267,7 +291,7 @@ TEMPLATE_TEST_CASE("[Hermitian]", "[MPSTNCuda_Expval]", float, double) {
             mps_state.applyOperations({{"Hadamard"}, {"CNOT"}, {"CNOT"}},
                                       {{0}, {0, 1}, {1, 2}},
                                       {{false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = HermitianObsT(mat, std::vector<std::size_t>{0});
             auto res = measure.expval(ob);
             CHECK(res == ZERO);
@@ -277,7 +301,7 @@ TEMPLATE_TEST_CASE("[Hermitian]", "[MPSTNCuda_Expval]", float, double) {
             mps_state.applyOperations(
                 {{"Hadamard"}, {"Hadamard"}, {"Hadamard"}}, {{0}, {1}, {2}},
                 {{false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = HermitianObsT(mat, {0});
             auto res = measure.expval(ob);
             CHECK(res == Approx(ONE));
@@ -293,7 +317,7 @@ TEMPLATE_TEST_CASE("[Hermitian]", "[MPSTNCuda_Expval]", float, double) {
                  {"Hadamard"}},
                 {{0}, {0}, {1}, {1}, {2}, {2}},
                 {{false}, {false}, {false}, {false}, {false}, {false}});
-            mps_state.get_final_state();
+            mps_state.append_mps_final_state();
             auto ob = HermitianObsT(mat, {0});
             auto res = measure.expval(ob);
             CHECK(res == -Approx(ONE));
@@ -303,9 +327,9 @@ TEMPLATE_TEST_CASE("[Hermitian]", "[MPSTNCuda_Expval]", float, double) {
 
 TEMPLATE_TEST_CASE("Test expectation value of TensorProdObs",
                    "[MPSTNCuda_Expval]", float, double) {
-    using StateTensorT = MPSTNCuda<TestType>;
-    using NamedObsT = NamedObsTNCuda<StateTensorT>;
-    using TensorProdObsT = TensorProdObsTNCuda<StateTensorT>;
+    using TensorNetT = MPSTNCuda<TestType>;
+    using NamedObsT = NamedObsTNCuda<TensorNetT>;
+    using TensorProdObsT = TensorProdObsTNCuda<TensorNetT>;
     auto ZERO = TestType(0);
     auto INVSQRT2 = TestType(0.707106781186547524401);
     SECTION("Using XZ") {
@@ -313,12 +337,12 @@ TEMPLATE_TEST_CASE("Test expectation value of TensorProdObs",
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
         mps_state.applyOperations({{"Hadamard"}, {"Hadamard"}, {"Hadamard"}},
                                   {{0}, {1}, {2}}, {{false}, {false}, {false}});
 
-        auto m = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto m = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto X0 =
             std::make_shared<NamedObsT>("PauliX", std::vector<std::size_t>{0});
@@ -335,12 +359,12 @@ TEMPLATE_TEST_CASE("Test expectation value of TensorProdObs",
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
         mps_state.applyOperations({{"Hadamard"}, {"Hadamard"}, {"Hadamard"}},
                                   {{0}, {1}, {2}}, {{false}, {false}, {false}});
 
-        auto m = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto m = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto H0 = std::make_shared<NamedObsT>("Hadamard",
                                               std::vector<std::size_t>{0});
@@ -357,21 +381,21 @@ TEMPLATE_TEST_CASE("Test expectation value of TensorProdObs",
 
 TEMPLATE_TEST_CASE("Test expectation value of HamiltonianObs",
                    "[MPSTNCuda_Expval]", float, double) {
-    using StateTensorT = MPSTNCuda<TestType>;
-    using NamedObsT = NamedObsTNCuda<StateTensorT>;
-    using HamiltonianObsT = HamiltonianTNCuda<StateTensorT>;
+    using TensorNetT = MPSTNCuda<TestType>;
+    using NamedObsT = NamedObsTNCuda<TensorNetT>;
+    using HamiltonianObsT = HamiltonianTNCuda<TensorNetT>;
     auto ONE = TestType(1);
     SECTION("Using XZ") {
         std::size_t bondDim = GENERATE(2);
         std::size_t num_qubits = 3;
         std::size_t maxBondDim = bondDim;
 
-        StateTensorT mps_state{num_qubits, maxBondDim};
+        TensorNetT mps_state{num_qubits, maxBondDim};
 
         mps_state.applyOperations({{"Hadamard"}, {"Hadamard"}, {"Hadamard"}},
                                   {{0}, {1}, {2}}, {{false}, {false}, {false}});
 
-        auto m = MeasurementsTNCuda<StateTensorT>(mps_state);
+        auto m = MeasurementsTNCuda<TensorNetT>(mps_state);
 
         auto X0 =
             std::make_shared<NamedObsT>("PauliX", std::vector<std::size_t>{0});